Refactored directory structure.
